+++ ember-intl/addon-test-support/setup-intl.ts
@@ -4,7 +4,7 @@ import type { IntlService } from 'ember-intl';
 import type { TOptions } from 'ember-intl/services/intl';
 import type { Formats, Translations } from 'ember-intl/types';
 
-import { missingMessage } from './-private/serialize-translation';
+import { missingMessage } from './-private/missing-message';
 import { addTranslations } from './add-translations';
 
 export interface IntlTestContext {
